Modify handling of large data file
The R code now looks for the data file locally, and if it's missing it will
download it and save it locally first. The Dockerfile now downloads the
data file into the correct location at image build time so the data doesn't
need to be downloaded every time the container is run.